+++ PortableCheckout/app/src/main/java/com/hackgt7/portablecheckout/ui/home/HomeFragment.java
@@ -1,21 +1,30 @@
 package com.hackgt7.portablecheckout.ui.home;
 
+import android.content.DialogInterface;
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
+import com.google.zxing.integration.android.IntentIntegrator;
+import com.google.zxing.integration.android.IntentResult;
+import com.hackgt7.portablecheckout.CaptureAct;
 import com.hackgt7.portablecheckout.R;
 
-public class HomeFragment extends Fragment {
+public class HomeFragment extends Fragment implements View.OnClickListener {
+    Button scanButton;
+    int scannerOut = -1;
 
     private HomeViewModel homeViewModel;
 
@@ -31,7 +40,47 @@ public class HomeFragment extends Fragment {
                 textView.setText(s);
             }
         });
+
+        this.scanButton = (Button) root.findViewById(R.id.scanBtn);
+        scanButton.setOnClickListener(this::onClick);
         return root;
 
+    }
+
+    @Override
+    public void onClick(View view) {
+        barScan();
+    }
+
+    private void barScan() {
+        IntentIntegrator integrator = new IntentIntegrator(getActivity());
+        integrator.setCaptureActivity(CaptureAct.class);
+        integrator.setOrientationLocked(false);
+        integrator.setDesiredBarcodeFormats(IntentIntegrator.PRODUCT_CODE_TYPES);
+        integrator.setPrompt("Scanning Bar Code");
+        integrator.initiateScan();
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        IntentResult result = IntentIntegrator.parseActivityResult(requestCode, resultCode, data);
+        if (result != null) {
+            if (result.getContents() != null) {
+                this.scannerOut = resultCode;
+                AlertDialog.Builder builder = new AlertDialog.Builder(getContext());
+                builder.setMessage(result.getContents());
+                builder.setTitle("Scanning Result");
+                builder.setPositiveButton("Scan Again",
+                        (DialogInterface dialog, int which) -> barScan())
+                        .setNegativeButton("finish",
+                                (DialogInterface dialog, int which) -> {});
+                AlertDialog dialog = builder.create();
+                dialog.show();
+            } else {
+                Toast.makeText(getContext(), "no results", Toast.LENGTH_SHORT);
+            }
+        } else {
+            super.onActivityResult(requestCode, resultCode, data);
+        }
     }
 }
